Refactor: Delete Person by object instead of name
Updated the `delete` function in the `CustomContentProvider` to delete a `Person` object by its properties instead of just by name. This allows for more flexibility in deleting specific entries.

Added update and delete functionality to the `MainActivity` in the `contentproviderread` module. This allows for modifying and removing entries in the database from the UI.

Modified the `deleteByName` function in the `PeopleDatabase` to accept a `Person` object and delete by its properties. This aligns with the change in the `CustomContentProvider`.